Remove old Contact and replace with  V2
