+++ cardigan/stories/components/Contact/Contact.stories.mdx
@@ -11,6 +11,4 @@ import * as stories from './Contact.stories.tsx';
 
   <Story story={stories.basic} />
 
-  <Story story={stories.contactv2} />
-
 <ArgsTable />
