--- common/views/components/Contact/Contact.V2.tsx
@@ -1,104 +1,0 @@
-import { FunctionComponent, ReactElement } from 'react';
-import styled from 'styled-components';
-import { font } from '@weco/common/utils/classnames';
-import { createScreenreaderLabel } from '@weco/common/utils/telephone-numbers';
-import Space from '../styled/Space';
-import Icon from '@weco/common/views/components/Icon/Icon';
-import { phone as phoneIcon, email as emailIcon } from '@weco/common/icons';
-
-const Wrapper = styled(Space).attrs({
-  h: { size: 'm', properties: ['padding-left'] },
-  className: 'body-text',
-})`
-  border-left: 5px solid ${props => props.theme.color('accent.turquoise')};
-`;
-
-const TitleWrapper = styled(Space).attrs({
-  v: { size: 's', properties: ['margin-bottom'] },
-})`
-  display: block;
-`;
-
-const Title = styled.span.attrs({ className: font('intb', 4) })``;
-
-const Subtitle = styled(Space).attrs({
-  as: 'span',
-  h: { size: 's', properties: ['margin-left'] },
-  className: font('intr', 4),
-})``;
-
-const PhoneNumber = styled.span.attrs({ className: font('intr', 4) })`
-  display: block;
-`;
-
-const WithIconWrapper = styled(Space).attrs({
-  v: { size: 's', properties: ['margin-bottom', 'column-gap'] },
-})`
-  display: flex;
-  align-items: center;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-
-  .icon {
-    margin-top: 5px;
-    width: 25px;
-    height: 25px;
-
-    ${props => props.theme.media('medium')`
-        width: 35px;
-        height: 35px;
-    `}
-  }
-`;
-
-export type Props = {
-  title: string;
-  subtitle: string | null;
-  phone: string | null;
-  email: string | null;
-};
-
-const ContactV2: FunctionComponent<Props> = ({
-  title,
-  subtitle,
-  phone,
-  email,
-}: Props): ReactElement => {
-  return (
-    <Wrapper>
-      <TitleWrapper>
-        <Title>{title}</Title>
-        {subtitle && <Subtitle>{subtitle}</Subtitle>}
-      </TitleWrapper>
-
-      {phone && (
-        <WithIconWrapper>
-          <Icon icon={phoneIcon} />
-          <div>
-            <span className="visually-hidden">
-              {createScreenreaderLabel(phone)}
-            </span>
-            <PhoneNumber aria-hidden="true">{phone}</PhoneNumber>
-          </div>
-        </WithIconWrapper>
-      )}
-
-      {email && (
-        <WithIconWrapper>
-          <Icon icon={emailIcon} />
-          <a
-            style={{ display: 'block' }}
-            className={font('intr', 4)}
-            href={`mailto:${email}`}
-          >
-            {email}
-          </a>
-        </WithIconWrapper>
-      )}
-    </Wrapper>
-  );
-};
-
-export default ContactV2;
