+++ common/views/components/Contact/Contact.tsx
@@ -2,9 +2,9 @@ import { FunctionComponent, ReactElement } from 'react';
 import styled from 'styled-components';
 import { font } from '@weco/common/utils/classnames';
 import { createScreenreaderLabel } from '@weco/common/utils/telephone-numbers';
-import { useToggles } from '@weco/common/server-data/Context';
 import Space from '../styled/Space';
-import ContactV2 from './Contact.V2';
+import Icon from '@weco/common/views/components/Icon/Icon';
+import { phone as phoneIcon, email as emailIcon } from '@weco/common/icons';
 
 const Wrapper = styled(Space).attrs({
   h: { size: 'm', properties: ['padding-left'] },
@@ -13,7 +13,9 @@ const Wrapper = styled(Space).attrs({
   border-left: 5px solid ${props => props.theme.color('accent.turquoise')};
 `;
 
-const TitleWrapper = styled.span`
+const TitleWrapper = styled(Space).attrs({
+  v: { size: 's', properties: ['margin-bottom'] },
+})`
   display: block;
 `;
 
@@ -29,6 +31,28 @@ const PhoneNumber = styled.span.attrs({ className: font('intr', 4) })`
   display: block;
 `;
 
+const WithIconWrapper = styled(Space).attrs({
+  v: { size: 's', properties: ['margin-bottom', 'column-gap'] },
+})`
+  display: flex;
+  align-items: center;
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+
+  .icon {
+    margin-top: 5px;
+    width: 25px;
+    height: 25px;
+
+    ${props => props.theme.media('medium')`
+        width: 35px;
+        height: 35px;
+    `}
+  }
+`;
+
 export type Props = {
   title: string;
   subtitle: string | null;
@@ -42,18 +66,6 @@ const Contact: FunctionComponent<Props> = ({
   phone,
   email,
 }: Props): ReactElement => {
-  const { visualStories } = useToggles();
-
-  if (visualStories)
-    return (
-      <ContactV2
-        title={title}
-        subtitle={subtitle}
-        phone={phone}
-        email={email}
-      />
-    );
-
   return (
     <Wrapper>
       <TitleWrapper>
@@ -62,20 +74,28 @@ const Contact: FunctionComponent<Props> = ({
       </TitleWrapper>
 
       {phone && (
-        <>
-          <span className="visually-hidden">
-            {createScreenreaderLabel(phone)}
-          </span>
-          <PhoneNumber aria-hidden="true">{phone}</PhoneNumber>
-        </>
+        <WithIconWrapper>
+          <Icon icon={phoneIcon} />
+          <div>
+            <span className="visually-hidden">
+              {createScreenreaderLabel(phone)}
+            </span>
+            <PhoneNumber aria-hidden="true">{phone}</PhoneNumber>
+          </div>
+        </WithIconWrapper>
       )}
 
       {email && (
-        <div>
-          <a className={font('intr', 4)} href={`mailto:${email}`}>
+        <WithIconWrapper>
+          <Icon icon={emailIcon} />
+          <a
+            style={{ display: 'block' }}
+            className={font('intr', 4)}
+            href={`mailto:${email}`}
+          >
             {email}
           </a>
-        </div>
+        </WithIconWrapper>
       )}
     </Wrapper>
   );
